Organize global event handlers differently
